[fix]: fix count_nonzero, randperm, index_put
- count_nonzero: add tl.abs(x) for floating inputs to handle -0.0 in
  both kernel variants
- randperm: add explicit .to(torch.int32) for sort/shuffle indices
  before transferring to device
- index_put: specify dtype=torch.int32 for sophgo backend in
  gen_indices to fix index dtype mismatch

+++ src/flag_gems/runtime/backend/_sophgo/ops/randperm.py
@@ -129,7 +129,7 @@ def sort_by_key(key, value, valid_bits, generator=None):
     key_cpu = key.cpu()
     sorted_key_cpu, sorted_indices_cpu = torch.sort(key_cpu)
 
-    sorted_indices = sorted_indices_cpu.to(key.device)
+    sorted_indices = sorted_indices_cpu.to(torch.int32).to(key.device)
 
     # ========== Step 2: Use TPU scatter to reorder values ==========
     # This is the main compute-intensive operation, executed on TPU.
@@ -170,7 +170,7 @@ def sort_by_key(key, value, valid_bits, generator=None):
         # Sort random keys (lightweight CPU operation)
         random_keys_cpu = random_keys.cpu()
         _, shuffle_indices_cpu = torch.sort(random_keys_cpu)
-        shuffle_indices = shuffle_indices_cpu.to(key.device)
+        shuffle_indices = shuffle_indices_cpu.to(torch.int32).to(key.device)
 
         # Apply shuffle on TPU
         sorted_value = sorted_value[shuffle_indices]
